refactor!: rename package to @agntn/archives
